Dawn Update Start:
Well, this will be a bit disastious since the Dawn library used in the zgpu is largely outdated, and since the
recent imgui update for zgui, this completely breaks the library when webgpu is used. This not only affect my
future project that uses zgui, but also the future of zig game dev. Thus, to save the repo, I am going to try
something on the zgui side so that:

1. I can find a working solution to update the dawn library
2. Document the solution for the future updates
3. If possible, let me see if it is possible to let us freely pick between webgpu and dawn, reduce the
   reliance on Google projects for people who prefer a Degoogled Solutions.

These are hugely difficult tasks and I am not sure if I can solve it due to my lack of understanding to gpu
architecture, but I guess this will be a great experience to learn about the library such that I will have
better ideas how to use dawn or zgpu in the future.