- Replace EnumerateFeatures with GetFeatures
Since this update, I will update the code frequently
since each code change will reflect the exact pr
from the DAWN and Webgpu library such that to proof that
each of the update is based on evidence from other
pr instead of just assuming things, to ensure that the
upgrade made correctly for any item that has a function
name change. If the function name retains the same,
but the parameters are changed, I won't refer the
particular pr since they are obvious changes.

This update has replaced WGPUAdapterEnumerateFeatures
with WGPUProcAdapterGetFeatures because of an pr made
in 2024:

- https://issues.chromium.org/issues/368672123
- https://github.com/webgpu-native/webgpu-headers/pull/348